+++ src/main/java/roomescape/service/dto/ReservationSlotRequest.java
@@ -1,0 +1,10 @@
+package roomescape.service.dto;
+
+import java.time.LocalDate;
+
+public record ReservationSlotRequest(
+        LocalDate date,
+        Long timeId,
+        Long themeId
+) {
+}
